// frontend/src/app/enums/category-type.enum.ts
export enum CategoryType {
    ONE_HUNDRED_PERCENT = 'ONE_HUNDRED_PERCENT',
    LOVED_GAME = 'LOVED_GAME',
    BAD_GAME = 'BAD_GAME',
    DOESNT_COUNT = 'DOESNT_COUNT',
    BUGGED_GAME = 'BUGGED_GAME',
    LONG_GAME = 'LONG_GAME',
    VERY_LONG_GAME = 'VERY_LONG_GAME',
    ULTRA_LONG_GAME = 'ULTRA_LONG_GAME',
    HARD_GAME = 'HARD_GAME',
    VERY_HARD_GAME = 'VERY_HARD_GAME',
    ULTRA_HARD_GAME = 'ULTRA_HARD_GAME',
}